requirements.txt
playwright==1.47.0
simhash==2.1.2
pyee==12.0.0
brotli==1.1.0
requests==2.32.3
beautifulsoup4==4.12.3
bs4==0.0.2
python-dotenv==1.0.0
lxml==4.9.3
joblib==1.4.0
zstandard==0.23.0
msgpack==1.1.0
psutil==6.0.0
async_timeout==4.0.3
aiodns==3.2.0
pulsar-client==3.1.0
coolname==2.2.0